+++ packages/plugins/PointerPosition/src/types.ts
@@ -1,0 +1,15 @@
+import {
+  PointerPositionConfiguration,
+  PointerPositionProjection,
+} from '@polar/lib-custom-types'
+
+export interface PointerPositionState {
+  selectedProjection: number
+  pointerPosition: number[]
+}
+
+export interface PointerPositionGetters extends PointerPositionState {
+  configuration: PointerPositionConfiguration
+  projections: Required<PointerPositionProjection>[]
+  coordinateString: string
+}