+++ packages/clients/snowbox/src/mapConfiguration.ts
@@ -325,4 +325,12 @@ export const mapConfiguration = {
       waitMs: 300,
     },
   },
+  pointerPosition: {
+    projections: [
+      { code: 'EPSG:4326' },
+      { code: 'EPSG:25832', decimals: 0 },
+      { code: 'EPSG:3857', decimals: 1 },
+      { code: 'EPSG:31467', decimals: 2 },
+    ],
+  },
 }