+++ packages/core/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## unpublished
 
+- Fix: Change `EPSG:31467` projection definition string to be usable for `proj` conversions.
+- Fix: List items that are currently highlighted according to vuetify (e.g. options in a select element) now by default have the primary color as outline to improve accessibility.
 - Chore: Add teardown hints for single page application to README file.
 
 ## 3.2.0